chore: update ButtonTab -> ButtonGroup
